requirements.txt
crewai>=0.28.0
langchain-openai>=0.1.0
PyMuPDF>=1.23.0
Flask>=2.3.0
Flask-Cors>=4.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
supabase>=2.0.0
pandas>=2.0.0
requests>=2.31.0
typing-extensions>=4.7.0
reportlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0